new icon and design for selection tool and few other small cleanups
