+++ src/component/menu/menucomponents/toolring.js
@@ -1,85 +1,35 @@
-import React, { useState } from "react"
+import React from "react"
 import { useSelector } from "react-redux"
 import { CgErase, CgController } from "react-icons/cg"
-import { BsPencil } from "react-icons/bs"
+import { BiSelection } from "react-icons/bi"
 import store from "../../../redux/store"
 import { SET_TYPE } from "../../../redux/slice/drawcontrol"
 import { toolType } from "../../../constants"
-import StylePicker from "./stylepicker"
+import PenTool from "./pentool"
 import ShapeTools from "./shapetools"
 
 export default function ToolRing() {
-    const [open, setOpen] = useState(false)
     const typeSelector = useSelector(
         (state) => state.drawControl.liveStroke.type
-    )
-    const colorSelector = useSelector(
-        (state) => state.drawControl.liveStroke.style.color
     )
 
     return (
         <>
-            <div className="session-dialog-div">
-                {typeSelector === toolType.PEN ? (
-                    <button
-                        type="button"
-                        style={{ color: colorSelector }}
-                        id="icon-button-active"
-                        onClick={() => setOpen(true)}>
-                        <BsPencil id="icon" />
-                    </button>
-                ) : (
-                    <button
-                        type="button"
-                        id="icon-button"
-                        onClick={() => {
-                            store.dispatch(SET_TYPE(toolType.PEN))
-                        }}>
-                        <BsPencil id="icon" />
-                    </button>
-                )}
-                {
-                    // Palette Popup
-                    open ? (
-                        <div className="popup">
-                            <div
-                                role="button"
-                                tabIndex="0"
-                                className="cover"
-                                onClick={() => setOpen(false)}
-                                onKeyPress={() => {}}
-                            />
-                            <StylePicker />
-                        </div>
-                    ) : null
-                }
-            </div>
-            {typeSelector === toolType.ERASER ? (
-                <button
-                    style={{ color: colorSelector }}
-                    type="button"
-                    id="icon-button-active"
-                    onClick={() => {
-                        store.dispatch(SET_TYPE(toolType.ERASER))
-                    }}>
-                    <CgErase id="icon" />
-                </button>
-            ) : (
-                <button
-                    type="button"
-                    id="icon-button"
-                    onClick={() => {
-                        store.dispatch(SET_TYPE(toolType.ERASER))
-                    }}>
-                    <CgErase id="icon" />
-                </button>
-            )}
+            <PenTool />
             <button
-                style={
-                    typeSelector === toolType.DRAG
-                        ? { color: colorSelector }
-                        : null
+                type="button"
+                id={
+                    typeSelector === toolType.ERASER
+                        ? "icon-button-active"
+                        : "icon-button"
                 }
+                onClick={() => {
+                    store.dispatch(SET_TYPE(toolType.ERASER))
+                }}>
+                <CgErase id="icon" />
+            </button>
+            <ShapeTools />
+            <button
                 type="button"
                 id={
                     typeSelector === toolType.DRAG
@@ -91,7 +41,6 @@ export default function ToolRing() {
                 }}>
                 <CgController id="icon" />
             </button>
-            <ShapeTools />
             <button
                 type="button"
                 id={
@@ -102,7 +51,7 @@ export default function ToolRing() {
                 onClick={() => {
                     store.dispatch(SET_TYPE(toolType.SELECT))
                 }}>
-                <CgController id="icon" />
+                <BiSelection id="icon" />
             </button>
         </>
     )
